:sparkles: feat: adiciona interatividade com o usuario ao mudar usuario e senha

+++ editar-perfil/editar-perfil.js
@@ -8,6 +8,12 @@ const novaSenhaInput = document.getElementById("nova-senha")
 const confirmaNovaSenhaInput = document.getElementById("confirma-nova-senha")
 const spanErroSenha = formSenha.querySelector(".mensagem-de-erro")
 
+const spanUsuario = document.getElementById("usuario")
+
+const nomeDeUsuario = JSON.parse(localStorage.getItem("dados")).usuario
+
+spanUsuario.innerText = nomeDeUsuario
+
 // * apaga mensagem de erro ao clicar novamente no input
 const todosInputs = document.querySelectorAll("input")
 for (const input of todosInputs) {
@@ -27,9 +33,21 @@ formUsuario.addEventListener("submit", (e) => {
     return
   }
 
+  const opcao = confirm(`Tem certeza que deseja mudar seu usuário para "${novoUsuario}"?`)
+
+  if (!opcao) {
+    alert("Mudança de usuário cancelada.")
+    e.target.reset()
+    return
+  }
+
   const dadosAntigos = JSON.parse(localStorage.getItem("dados"))
   const dadosNovos = { ...dadosAntigos, usuario: novoUsuario }
   localStorage.setItem("dados", JSON.stringify(dadosNovos))
+
+  alert("Mudança de usuário concluída com sucesso!")
+
+  location.reload()
 })
 
 formSenha.addEventListener("submit", (e) => {
@@ -51,6 +69,17 @@ formSenha.addEventListener("submit", (e) => {
     return
   }
 
+  const opcao = confirm(`Tem certeza que deseja mudar sua senha?`)
+
+  if (!opcao) {
+    alert("Mudança de senha cancelada.")
+    e.target.reset()
+    return
+  }
+
   const dadosNovos = { ...dadosAntigos, senha: novaSenha }
   localStorage.setItem("dados", JSON.stringify(dadosNovos))
+
+  alert("Mudança de senha concluída com sucesso!")
+  e.target.reset()
 })
